fix(colima,brew): setup colima to autostart using `brew services`
Homebrew needed to be fed the extra env variable `COLIMA_HOME` for it to
know that it should pick things from `XDG_CONFIG_HOME/colima`. Since
that was not available, it picked up `$HOME/.colima` instead; which
causes issues with colima since in my normal ZSH environment, that
variable is defined, so things are looking in two different directories
to find sockets that do not exist.

Signed-off-by: Amit Singh <amitds1997@gmail.com>